+++ consensus/types/src/lib.rs
@@ -102,7 +102,6 @@ pub mod blob_sidecar;
 pub mod blobs_sidecar;
 pub mod signed_blob;
 pub mod signed_block_and_blobs;
-pub mod signed_block_contents;
 pub mod transaction;
 
 use ethereum_types::{H160, H256};
@@ -185,10 +184,8 @@ pub use crate::signed_beacon_block::{
 pub use crate::signed_beacon_block_header::SignedBeaconBlockHeader;
 pub use crate::signed_blob::*;
 pub use crate::signed_block_and_blobs::{
-    SignedBeaconBlockAndBlobSidecars, SignedBeaconBlockAndBlobsSidecar,
-    SignedBeaconBlockAndBlobsSidecarDecode,
+    SignedBeaconBlockAndBlobsSidecar, SignedBeaconBlockAndBlobsSidecarDecode,
 };
-pub use crate::signed_block_contents::SignedBlockContents;
 pub use crate::signed_bls_to_execution_change::SignedBlsToExecutionChange;
 pub use crate::signed_contribution_and_proof::SignedContributionAndProof;
 pub use crate::signed_voluntary_exit::SignedVoluntaryExit;
